+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/test/java/com/tencent/cloud/polaris/circuitbreaker/PolarisCircuitBreakerMockServerTest.java
@@ -30,6 +30,7 @@ import java.util.stream.Collectors;
 
 import com.google.protobuf.util.JsonFormat;
 import com.tencent.cloud.common.util.ApplicationContextAwareUtils;
+import com.tencent.cloud.polaris.context.PolarisSDKContextManager;
 import com.tencent.polaris.api.config.Configuration;
 import com.tencent.polaris.api.core.ConsumerAPI;
 import com.tencent.polaris.api.pojo.ServiceKey;
@@ -76,7 +77,7 @@ public class PolarisCircuitBreakerMockServerTest {
 				.thenReturn(NAMESPACE_TEST);
 		mockedApplicationContextAwareUtils.when(() -> ApplicationContextAwareUtils.getProperties("spring.cloud.polaris.service"))
 				.thenReturn(SERVICE_CIRCUIT_BREAKER);
-
+		PolarisSDKContextManager.innerDestroy();
 		namingServer = NamingServer.startNamingServer(-1);
 		System.setProperty(SERVER_ADDRESS_ENV, String.format("127.0.0.1:%d", namingServer.getPort()));
 
@@ -99,7 +100,9 @@ public class PolarisCircuitBreakerMockServerTest {
 		if (null != namingServer) {
 			namingServer.terminate();
 		}
-		mockedApplicationContextAwareUtils.close();
+		if (null != mockedApplicationContextAwareUtils) {
+			mockedApplicationContextAwareUtils.close();
+		}
 	}
 
 	@Test